requirements.txt

### 2. requirements.txt
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.18.0
peft>=0.3.0
torch>=1.12.0
fastapi>=0.78.0
uvicorn>=0.18.0
pyyaml>=6.0
pytest>=7.0
